+++ src/irmin-pack/mem/indexable.ml
@@ -60,9 +60,8 @@ module Maker (K : Irmin.Hash.S) = struct
       (Val : Irmin_pack.Pack_value.S with type hash := K.t and type key := K.t) =
   struct
     (* TODO(craigfe): We could use the keys to skip traversal of the map on
-       lookup, but this must not introduce false-positives under [clear].
-
-       See https://github.com/mirage/irmin/pull/1389#discussion_r693789934. *)
+       lookup. This wasn't done originally due to complications with implementing
+       the [clear] function, but this has since been removed. (See #1794.) *)
     module Key = Irmin.Key.Of_hash (K)
 
     module KMap = Map.Make (struct
@@ -81,11 +80,6 @@ module Maker (K : Irmin.Hash.S) = struct
     let instances = Pool.create ~alloc:(fun name -> { name; t = KMap.empty })
     let v name = Lwt.return (Pool.take instances name)
     let equal_key = Irmin.Type.(unstage (equal K.t))
-
-    let clear t =
-      [%log.debug "clear"];
-      t.t <- KMap.empty;
-      Lwt.return_unit
 
     let close t =
       [%log.debug "close"];