+++ src/irmin-test/store.ml
@@ -1872,40 +1872,6 @@ module Make (S : Generic_key) = struct
     in
     run x test
 
-  let test_reexport_node x () =
-    let test repo1 =
-      let* tree = S.Tree.add (S.Tree.empty ()) [ "foo"; "a" ] "a" in
-      let* _ =
-        S.Backend.Repo.batch repo1 (fun c n _ -> S.save_tree repo1 c n tree)
-      in
-      let* () = B.Repo.close repo1 in
-      (* Re-export the same tree using a different repo. *)
-      let* repo2 = S.Repo.v x.config in
-      let* _ =
-        check_raises_lwt "re-export tree from another repo"
-          (Failure "Can't export the node key from another repo") (fun () ->
-            S.Backend.Repo.batch repo2 (fun c n _ -> S.save_tree repo2 c n tree))
-      in
-      let* () = B.Repo.close repo2 in
-      (* Re-export a fresh tree using a different repo. *)
-      let* repo2 = S.Repo.v x.config in
-      let* tree = S.Tree.add (S.Tree.empty ()) [ "foo"; "a" ] "a" in
-      let _ = S.Tree.hash tree in
-      let* c1 = S.Tree.get_tree tree [ "foo" ] in
-      let* _ =
-        S.Backend.Repo.batch repo2 (fun c n _ -> S.save_tree repo2 c n c1)
-      in
-      let* () =
-        match S.Tree.destruct c1 with
-        | `Contents _ -> Alcotest.fail "got `Contents, expected `Node"
-        | `Node node ->
-            let* _v = S.to_backend_node node in
-            Lwt.return_unit
-      in
-      B.Repo.close repo2
-    in
-    run x test
-
   module Sync = Irmin.Sync.Make (S)
 
   let test_sync x () =
@@ -2387,59 +2353,6 @@ module Make (S : Generic_key) = struct
     in
     run x test
 
-  let test_clear x () =
-    let test repo =
-      let vt = v repo and b = b repo and ct = ct repo and n = n repo in
-      let check_none msg =
-        Alcotest.(check (testable Fmt.(option (const string "<value>")) ( = )))
-          msg None
-      in
-      let check_val = check (T.option S.contents_t) in
-      let check_commit = check (T.option @@ S.commit_t repo) in
-      let* h2 = kv2 ~repo in
-      B.Contents.clear vt >>= fun () ->
-      let* () =
-        B.Contents.find vt h2 >|= check_val "v2 after clear is not found" None
-      in
-      let* h2 = kv2 ~repo in
-      let* () =
-        B.Contents.find vt h2 >|= check_val "add v2 again after clear" (Some v2)
-      in
-      let* h1 = r1 ~repo in
-      let* n1 = n1 ~repo in
-      S.Branch.set repo b1 h1 >>= fun () ->
-      let* () =
-        S.Branch.find repo b1
-        >|= check_commit "r1 before clear is found" (Some h1)
-      in
-      let* () =
-        Lwt.join [ B.Branch.clear b; B.Commit.clear ct; B.Node.clear n ]
-      in
-      let* () =
-        S.Branch.find repo b1
-        >|= check_commit "r1 after clear is not found" None
-      in
-      let* () =
-        B.Commit.find ct (S.Commit.key h1)
-        >|= check_none "after clear commit is not found"
-      in
-      let* () =
-        B.Node.find n n1 >|= check_none "after clear node is not found"
-      in
-      let* h2 = kv2 ~repo in
-      B.Contents.clear vt >>= fun () ->
-      let* () =
-        B.Contents.find vt h2 >|= check_none "v2 after clear is not found"
-      in
-      r1 ~repo >>= S.Branch.set repo b1 >>= fun () ->
-      B.Branch.clear b >>= fun () ->
-      let* () =
-        S.Branch.find repo b1 >|= check_commit "cleared twice r1" None
-      in
-      B.Repo.close repo
-    in
-    run x test
-
   let test_pre_hash_collisions x () =
     let pre_hash_of ty =
       let f = Irmin.Type.(pre_hash ty |> unstage) in
@@ -2550,9 +2463,7 @@ let suite (speed, x) =
        ("Shallow objects", speed, T.test_shallow_objects x);
        ("Closure with disconnected commits", speed, T.test_closure x);
        ("Prehash collisions", speed, T.test_pre_hash_collisions x);
-       ("Re-export node", speed, T.test_reexport_node x);
      ]
-    @ when_ x.clear_supported [ ("Clear", speed, T.test_clear x) ]
     @ when_ x.import_supported
         [
           ("Basic operations on slices", speed, T.test_slice x);